requirements.txt
boto3==1.34.149
botocore==1.34.149
certifi==2024.7.4
charset-normalizer==3.3.2
Faker==26.0.0
idna==3.7
jmespath==1.0.1
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
